refact: move method to public.go and private.go
