+++ internal/fswatcher/private.go
