+++ internal/fswatcher/public.go
