+++ internal/nftman/private_release.go
@@ -7,10 +7,6 @@ import (
 	"github.com/google/nftables/expr"
 )
 
-func (t *Table) DumpNFTableRules() {
-	return
-}
-
 func addDebugCounter(exprs []expr.Any) []expr.Any {
 	return exprs
 }
