+++ internal/nftman/public_release.go
@@ -1,0 +1,8 @@
+//go:build !debug
+// +build !debug
+
+package nftman
+
+func (t *Table) DumpNFTableRules() {
+	return
+}
